complete sign-up for customer, view cart, add to cart, remove from cart, create order

+++ src/main/java/com/example/electriccomponentsshop/entities/OrderItem.java
@@ -6,6 +6,8 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
+
 @Entity
 @Table(name="Order_Items")
 @NoArgsConstructor
@@ -15,13 +17,18 @@ import javax.persistence.*;
 public class OrderItem {
     @EmbeddedId
     private OrderItemId orderItemId;
-    private Double unitPrice;
-    private Integer quantity;
-    private Double subTotal;
+
+    private BigDecimal unitPrice;
+
+    private BigDecimal quantity;
+
+    private BigDecimal subTotal;
+
     @ManyToOne(fetch = FetchType.EAGER)
     @MapsId("orderId")
     @JoinColumn(name="order_id")
     private Order order;
+
     @ManyToOne(fetch = FetchType.EAGER)
     @MapsId("productId")
     @JoinColumn(name="product_id")
